Add changePriceColor to helper, Update CoinInfo style
